rename metadata as per PR comments
