+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/assumptions.rs
@@ -4,7 +4,7 @@
 
 use crate::gotoc::cbmc::goto_program::{Expr, Location, Stmt, Symbol, Type};
 use crate::gotoc::mir_to_goto::codegen::typ::tuple_fld;
-use crate::gotoc::mir_to_goto::metadata::GotocCtx;
+use crate::gotoc::mir_to_goto::GotocCtx;
 use rustc_middle::mir::interpret::{ConstValue, Scalar};
 use rustc_middle::ty;
 use rustc_middle::ty::ScalarInt;
