+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/codegen/rvalue.rs
@@ -5,8 +5,8 @@ use crate::btree_string_map;
 use crate::gotoc::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Symbol, Type};
 use crate::gotoc::cbmc::utils::{aggr_name, BUG_REPORT_URL};
 use crate::gotoc::cbmc::MachineModel;
-use crate::gotoc::mir_to_goto::metadata::*;
 use crate::gotoc::mir_to_goto::utils::{dynamic_fat_ptr, slice_fat_ptr};
+use crate::gotoc::mir_to_goto::GotocCtx;
 use num::bigint::BigInt;
 use rustc_middle::mir::{AggregateKind, BinOp, CastKind, NullOp, Operand, Place, Rvalue, UnOp};
 use rustc_middle::ty::adjustment::PointerCast;
