+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/hooks.rs
@@ -9,9 +9,9 @@
 //! this module addresses this issue.
 
 use crate::gotoc::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Symbol, Type};
-use crate::gotoc::mir_to_goto::metadata::GotocCtx;
 use crate::gotoc::mir_to_goto::stubs::hash_map_stub::HashMapStub;
 use crate::gotoc::mir_to_goto::stubs::vec_stub::VecStub;
+use crate::gotoc::mir_to_goto::GotocCtx;
 use rustc_hir::definitions::DefPathDataName;
 use rustc_middle::mir::{BasicBlock, Place};
 use rustc_middle::ty::print::with_no_trimmed_paths;
